requirements.txt
flask==3.1.1
flask-login==0.6.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
werkzeug==3.1.3
gunicorn==23.0.0
